python/TODO_name/requirements.txt
unipath
Django==1.8.7
django-registration
django-bootstrap3
django-finalware